@@ -3,9 +3,9 @@ package in.codeshuffle.examplelinkpreview.ui.fragment.demoscreen;
 import android.content.Context;
 import android.graphics.drawable.Drawable;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -22,7 +22,6 @@ import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.Unbinder;
 import in.codeshuffle.examplelinkpreview.R;
-import in.codeshuffle.examplelinkpreview.ui.activity.main.MainActivity;
 import in.codeshuffle.examplelinkpreview.util.GlideApp;
 import in.codeshuffle.linkpreviewedittext.LinkInfo;
 import in.codeshuffle.linkpreviewedittext.LinkPreviewEditText;
